fixing delete button and twitchwebsocket models
