+++ app/src/main/java/com/example/clicker/network/models/websockets/DataClases.kt
@@ -3,7 +3,7 @@ package com.example.clicker.network.models.websockets
 import androidx.compose.runtime.Immutable
 import com.example.clicker.network.websockets.EmoteInText
 import com.example.clicker.network.websockets.MessageToken
-import com.example.clicker.network.websockets.MessageType
+import com.example.clicker.network.websockets.models.MessageType
 
 /**
  * Represents the state of the logged in User
