+++ app/src/main/java/com/example/clicker/presentation/stream/StreamView.kt
@@ -28,6 +28,7 @@ import com.example.clicker.presentation.modView.ListTitleValue
 import com.example.clicker.presentation.modView.ModViewViewModel
 import com.example.clicker.presentation.modView.slowModeListImmutable
 import com.example.clicker.presentation.modView.followerModeListImmutable
+import com.example.clicker.presentation.stream.models.AdvancedChatSettings
 
 import com.example.clicker.presentation.stream.views.TestingNewBottomModal
 import com.example.clicker.presentation.stream.views.chat.BadgeSlider
