+++ app/src/main/java/com/example/clicker/util/objectMothers/IndividualAutoModSettingsDataObjectMother.kt
@@ -1,8 +1,7 @@
 package com.example.clicker.util.objectMothers
 
 import com.example.clicker.network.models.twitchStream.IndividualAutoModSettings
-import com.example.clicker.network.models.websockets.TwitchUserData
-import com.example.clicker.network.websockets.MessageType
+
 
 class IndividualAutoModSettingsDataObjectMother private constructor() {
