+++ app/src/main/java/com/example/clicker/util/objectMothers/TwitchUserDataObjectMother.kt
@@ -1,8 +1,8 @@
 package com.example.clicker.util.objectMothers
 
-import com.example.clicker.network.websockets.MessageType
 import com.example.clicker.network.models.websockets.TwitchUserData
 import com.example.clicker.network.websockets.MessageToken
+import com.example.clicker.network.websockets.models.MessageType
 
 class TwitchUserDataObjectMother private constructor() {
